Java OOPS For-Each Loop Practice

+++ Java OOP Inheritance/Developer.java
@@ -2,7 +2,7 @@ public class Developer extends Employee{
 
     private String programmingLanguage;
 
-    public Developer(String name, int salary, String programmingLanguage) {
+    public Developer(String name, double salary, String programmingLanguage) {
         super(name, salary);
         this.programmingLanguage = programmingLanguage;
     }
