+++ Java OOP Inheritance/Manager.java
@@ -2,7 +2,7 @@ public class Manager extends Employee {
 
     private int teamSize;
     
-    public Manager(String name, int salary, int teamSize) {
+    public Manager(String name, double salary, int teamSize) {
         super(name, salary);
         this.teamSize = teamSize;
     }
